requirements.txt
llama-cpp-python
spacy

wikipedia-api
torch
requests
transformers
SPARQLWrapper
